refactor(event-stream): extract RT_FLOW projection to rt_flow.rs
Pure code motion for #6235. Move the two RT_FLOW projection emitters out
of the EventStreamWorkerHandle impl in mod.rs into a dedicated rt_flow.rs
submodule as a second impl block: emit_session_close_rt_flow (SESSION_CLOSE
type-14 frame, #2460/#2465/#2853/#3395) and emit_session_create_rt_flow
(SESSION_CREATE type-15 frame, #2508/#2615). Both still route through the
shared per-kind rate-limiter + queue-budget path (try_emit_dataplane_frame
in producer.rs) and anchor monotonic instants to wall-clock via clock.rs.

Moved verbatim (no logic or signature change); the pub(crate) methods
resolve at their call sites (afxdp/session_delta.rs) unchanged. NS_PER_SEC
is widened to pub(super) and re-imported so the rt_flow test module
(which reaches it via `use super::*`) resolves it. The orphaned I/O-thread
section comment is removed and the README Files section is updated for the
new submodule layout.

This completes the #6235 split: mod.rs drops from 2001 to 733 lines
(facade + shared state + producer core + bootstrap); the transport /
producer / connection / replay / drain / control / budget / clock /
backlog / rt_flow concerns each own a cohesive submodule under the
2000-LOC modularity threshold. All 4 #[inline] attributes are preserved
(0 lost); no #[cold] existed. Full cargo test --release: 4214 passed, 0
failed, 2 ignored.

Closes #6235
